+++ RuntimeEditor/EditorHelper.cs
@@ -72,7 +72,7 @@ namespace Hananoki {
 				var texture = p.GetProperty<Texture>( "texture" );
 				p.gameObject.name = $"{prefix}{texture.name}";
 			}
-			else if( p.GetType().IsSubclassOf( UnityTypes.TMPro_TMP_Text )  ) {
+			else if( p.GetType().IsSubclassOf( UnityTypes.TMPro_TMP_Text ) ) {
 				var text = p.GetProperty<string>( "text" ); ;
 				Regex re = new Regex( "<.*?>", RegexOptions.Singleline );
 				text = re.Replace( text, "" );
@@ -903,120 +903,6 @@ namespace Hananoki {
 
 
 
-	/// <summary>
-	/// 
-	/// </summary>
-	public static class Icon {
-		static Dictionary<string, Texture2D> s_iconCache;
-
-		/// <summary>
-		/// 
-		/// </summary>
-		static Icon() {
-			s_iconCache = new Dictionary<string, Texture2D>();
-		}
-
-
-		public static Texture2D Get( string lightskin, string darkskin ) {
-			if( EditorGUIUtility.isProSkin ) {
-				return Get( darkskin );
-			}
-			return Get( lightskin );
-		}
-
-		public static string GetBuiltinPath( string path ) {
-			var paths = path.Split( '/' );
-			if( EditorGUIUtility.isProSkin ) {
-				paths[ paths.Length - 1 ] = "d_" + paths[ paths.Length - 1 ];
-			}
-
-			return string.Join( "/", paths );
-		}
-
-		public static Texture2D GetBuiltin( string path ) {
-			if( EditorGUIUtility.isProSkin ) {
-				var icon = Get( GetBuiltinPath( path ) );
-				if( icon != null ) return icon;
-			}
-			return Get( path );
-		}
-
-
-		public static Texture2D Get( string name ) {
-			if( s_iconCache.ContainsKey( name ) ) {
-				return s_iconCache[ name ];
-			}
-			if( name.IsEmpty() ) return null;
-
-			Texture2D icon;
-
-			try {
-				icon = EditorGUIUtility.FindTexture( name ) as Texture2D;
-				if( icon != null ) {
-					s_iconCache.Add( name, icon );
-					return icon;
-				}
-			}
-			catch( Exception ) {
-			}
-
-			try {
-				Texture2D iconz = EditorGUIUtility.Load( name ) as Texture2D;
-				//var iconz = EditorGUIUtility.IconContent( name ).image as Texture2D;
-				if( iconz != null ) {
-					s_iconCache.Add( name, iconz );
-					return iconz;
-				}
-			}
-			catch( Exception ) {
-			}
-
-			try {
-				Texture2D iconz = EditorGUIUtility.Load( "icons/" + name + ".png" ) as Texture2D;
-				//var iconz = EditorGUIUtility.IconContent( name ).image as Texture2D;
-				if( iconz != null ) {
-					s_iconCache.Add( name, iconz );
-					return iconz;
-				}
-			}
-			catch( Exception ) {
-			}
-
-			try {
-				Texture2D iconz = EditorGUIUtility.Load( name ) as Texture2D;
-				//var iconz = EditorGUIUtility.IconContent( name ).image as Texture2D;
-				if( iconz != null ) {
-					s_iconCache.Add( name, iconz );
-					return iconz;
-				}
-			}
-			catch( Exception ) {
-			}
-
-			var a = AssetDatabase.FindAssets( "Icons" );
-			foreach( var b in a ) {
-				var v = AssetDatabase.GUIDToAssetPath( b );
-				var vv = Path.GetExtension( v );
-				if( !string.IsNullOrEmpty( vv ) ) continue;
-
-				icon = AssetDatabase.LoadAssetAtPath<Texture2D>( v + "/" + name + ".png" );
-				if( icon != null ) {
-					//UnityEngine.Debug.Log( "m_iconCache: add: " + name );
-					s_iconCache.Add( name, icon );
-					return icon;
-				}
-			}
-
-			{
-				var lst = Resources.FindObjectsOfTypeAll<Texture2D>().Where( x => x.name.Contains( name ) ).ToArray();
-				if( 0 < lst.Length ) {
-					s_iconCache.Add( name, lst[ 0 ] );
-					return lst[ 0 ];
-				}
-			}
-			return null;
-		}
-	}
 
 
 	public static class B64 {